requirements.txt
ccxt
pandas
python-dotenv
aiohttp
pandas_ta
numpy
pyyaml
yfinance